requirements.txt
Flask==3.0.3
gunicorn==23.0.0
PyJWT==2.9.0
requests==2.32.3
scikit-learn==1.7.2
SQLAlchemy==2.0.36
psycopg[binary]>=3.2,<3.3
